requirements.txt
Django==3.2.22
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==0.21.1
djangorestframework-simplejwt==4.7.1